Added dynamic fee tx type and tx fields
