+++ blockchain/blockchain.go
@@ -1091,7 +1091,7 @@ func (b *Blockchain) ReadTxLookup(hash types.Hash) (types.Hash, bool) {
 // return error if the invalid signature found
 func (b *Blockchain) recoverFromFieldsInBlock(block *types.Block) error {
 	for _, tx := range block.Transactions {
-		if tx.From != types.ZeroAddress || tx.IsStateTx() {
+		if tx.From != types.ZeroAddress || tx.Type == types.StateTx {
 			continue
 		}
 
@@ -1112,7 +1112,7 @@ func (b *Blockchain) recoverFromFieldsInTransactions(transactions []*types.Trans
 	updated := false
 
 	for _, tx := range transactions {
-		if tx.From != types.ZeroAddress || tx.IsStateTx() {
+		if tx.From != types.ZeroAddress || tx.Type == types.StateTx {
 			continue
 		}
